add access control

+++ lib/DeleteAction.php
@@ -29,6 +29,15 @@ class DeleteAction extends BaseAction
             throw new NotFoundHttpException(Yii::t('app', 'The requested page does not exist.'));
         }
 
+        /**
+         * running action accessControl to check whether user has priviledges to run action
+         */
+        $passed = $this->accessControlFilter($model);
+
+        if ($passed === FALSE) {
+            return $this->fallbackPage($model);
+        }
+
         try {
             $model->delete();
         } catch (\Exception $e) {
